Add inline directive
