+++ src/ert/config/parsing/schema_item_type.py
@@ -9,6 +9,10 @@ class SchemaItemType(StrEnum):
     POSITIVE_FLOAT = "POSITIVE_FLOAT"
     PATH = "PATH"
     EXISTING_PATH = "EXISTING_PATH"
+    # EXISTING_PATH_INLINE is a directive to the
+    # schema validation to inline the contents of
+    # the file.
+    EXISTING_PATH_INLINE = "EXISTING_PATH_INLINE"
     BOOL = "BOOL"
     CONFIG = "CONFIG"
     BYTESIZE = "BYTESIZE"
